@@ -6,21 +6,25 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  form_validation: ^1.0.2
-  google_fonts: ^1.1.0
-  intl: ^0.16.1
+  form_validation: ^2.0.0
+  google_fonts: ^2.0.0
+  intl: ^0.17.0
   json_theme:
     path: ../
-  meta: ^1.2.4
+  meta: ^1.3.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
+dependency_overrides:
+  json_schema2:
+    path: ../../json_schema2
 
 flutter:
   uses-material-design: true
